Restructuring the benchmarks script, so a "Small|Large" keys can be used to perform a type of benchmark. Added recipe examples to show how JSON keys are used. Modified logic utilisation scripts to gather missing information that is used in new report format, as well as generate a CSV file with such information. Modified makefiles to execute either Small or Large benchmarks based on JSON keys.
