+++ benchmarks.py
@@ -19,50 +19,55 @@ def execute_command(cmd):
     if (process != 0):
         logging.error("An error occurred while running: {}".format(cmd))
 
-def execute_json(data):
+def run_main(runtype, **recipeFile):
+    command=('{} make -j$(nproc) DIRECTIVE={} DIR={}'.format(runtype, recipeFile['Directive'], recipeFile['Report']))
+    if (recipeFile["Tool"] == "vivado"):
+        logging.info("Moving to Vivado directory to store results. Executing command {}".format(command))
+        os.chdir(vivado_makefile)
+        execute_command(command+"&")
+        os.chdir(sys.path[0])
+    elif (recipeFile["Tool"] == "yosys"):
+        command=(command+' YSARGS="{}"'.format(recipeFile["YosysArgs"]))
+        logging.info("Moving to Yosys directory to store results. Executing command {}".format(command))
+        os.chdir(yosys_abc9_makefile)
+        execute_command(command+"&")
+        os.chdir(sys.path[0])
+    elif (recipeFile["Tool"] == "ise"):
+        logging.info("Moving to ISE directory to store results. Executing command {}".format(command))
+        os.chdir(ise_makefile)
+        execute_command(command+"&")
+        os.chdir(sys.path[0])
+    elif (recipeFile["Tool"] == "yosys_s6"):
+        command=(command+' YSARGS="{}"'.format(recipeFile["YosysArgs"]))
+        logging.info("Moving to Yosys S6 directory to store results. Executing command {}".format(command))
+        os.chdir(yosys_s6_makefile)
+        execute_command(command+"&")
+        os.chdir(sys.path[0])
+    else:
+        logging.warning("No process executed. Review the JSON recipe file")
+ 
+def execute_json2(data):
     """
     This function parses a JSON file
-    to get the information of tool for synthesis (yosys, vivado, ise)
+    to get the information of tool for synthesis (yosys, vivado, ise),
     directives and possibly custom flow (TODO)
     """
     with open (data, 'r') as f:
         recipeFile = json.load(f)
+        logging.info("Keys are {}".format(recipeFile.keys()))
 
-        for tool in recipeFile:
-            logging.info("Tool selected for running is: {}".format(tool['Tool']))
-
-        for directive in recipeFile:
-            logging.info("Directive to perform benchmark is: {}".format(directive["Directive"]))
-
-        for items in range(len(recipeFile)):
-            command=('make -j$(nproc) DIRECTIVE={} DIR={}'.format(recipeFile[items]["Directive"], recipeFile[items]["Report"]))
-            if (recipeFile[items]["Tool"] == "vivado"):
-                logging.info("Moving to Vivado directory to store results")
-                os.chdir(vivado_makefile)
-                logging.info("Executing Vivado Makefile")
-                execute_command(command)
-                os.chdir(sys.path[0])
-            elif (recipeFile[items]["Tool"] == "yosys"):
-                command=(command+' YSARGS="{}"&'.format(recipeFile[items]["YosysArgs"]))
-                logging.info("Moving to Yosys directory to store results. Executing command {}".format(command))
-                os.chdir(yosys_abc9_makefile)
-                logging.info("Executing Yosys Makefile")
-                execute_command(command)
-                os.chdir(sys.path[0])
-            elif (recipeFile[items]["Tool"] == "ise"):
-                logging.info("Moving to ISE directory to store results")
-                os.chdir(ise_makefile)
-                logging.info("Executing ISE Makefile")
-                logging.info("Command is {}".format(command))
-                execute_command(command)
-            elif (recipeFile[items]["Tool"] == "yosys_s6"):
-                logging.info("Moving to Yosys S6 directory to store results")
-                os.chdir(yosys_s6_makefile)
-                logging.info("Executing Yosys Makefile")
-                logging.info("Command is {}".format(command))
-                execute_command(command)
+        for key in recipeFile.keys():
+            logging.info("Key is {}".format(key))
+            if key in ('Large', 'Small'):
+                runtype=("{}=1".format(key.upper()))
+                for values in recipeFile[key]:
+                    logging.info("Tool selected for running is: {}".format(values['Tool']))
+                    logging.info("Directive to perform benchmark is: {}".format(values['Directive']))
+                    logging.info("Type of benchmark is: {}".format(key))
+                    run_main(runtype, **values)
             else:
-                logging.warning("No process executed. Review the JSON recipe file")
+                raise Exception('Value selected for benchmark type is wrong: {}. It shoule be either Small or Large. Please review the JSON file'.format(key))
+
 """ 
 Main
 """
@@ -72,5 +77,6 @@ if __name__=="__main__":
                         help="Recipe in JSON format for Benchmarks",
                         type=argparse.FileType('r'))
     arguments=parser.parse_args()
-    execute_json(arguments.recipe[0].name)
+    execute_json2(arguments.recipe[0].name)
+
