+++ ise_min_period/ise-14.7/Makefile
@@ -7,10 +7,11 @@ TARGET=$(DIR)
 TARGETS = $(patsubst %,tab_%_$(DEVICE)_$(GRADE)/results.txt,$(basename $(notdir $(BENCHMARKS))))
 
 all: $(TARGETS)
-	cd $(TARGET) && ../../../scripts/ise_min_period_table.sh > results_$(DEVICE)_$(DIR).csv
+	cd $(DIR) && ../../../scripts/ise_min_period_table.sh > results_$(DEVICE)_$(DIR).csv
 
 tab_%_$(DEVICE)_$(GRADE)/results.txt:
-	../../scripts/ise_min_period.sh ../../*/benchmarks_*/*/$*.{v.gz,v,vhdl} $(DEVICE) $(GRADE) --rdir $(DIR)
+	f=(../../*/benchmarks_*/*/$*.{v.gz,v,vhdl}); \
+	../../scripts/ise_min_period.sh $${f[0]} $(DEVICE) $(GRADE) --rdir $(DIR)
 
 clean:
-	 rm -r tab_*
+	 rm -r $(DIR)/tab_*
