+++ recipes/full_vivado_yosys_example_recipe.json
@@ -1,0 +1,30 @@
+{
+  "Large": [
+    {
+      "Tool": "yosys",
+      "Directive": "Default",
+      "Report": "yosys_large_default",
+      "YosysArgs": "-abc9"
+    },
+    {
+      "Tool": "vivado",
+      "Directive": "Default",
+      "Report": "vivado_large_default",
+      "YosysArgs": "-abc9"
+    }
+  ],
+  "Small": [
+    {
+      "Tool": "vivado",
+      "Directive": "Default",
+      "Report": "vivado_small_default",
+      "YosysArgs": "-abc9"
+    },
+    {
+      "Tool": "yosys",
+      "Directive": "Default",
+      "Report": "yosys_small_default",
+      "YosysArgs": "-abc9"
+    }
+  ]
+}
