+++ recipes/small_vivado_yosys_example_recipe.json
@@ -1,0 +1,15 @@
+{
+  "Small": [
+    {
+      "Tool": "vivado",
+      "Directive": "Default",
+      "Report": "vivado_small_default"
+    },
+    {
+      "Tool": "yosys",
+      "Directive": "Default",
+      "Report": "yosys_small_default",
+      "YosysArgs": "-abc9 -flatten"
+    }
+  ]
+}
