--- test.json
@@ -1,15 +1,0 @@
-{
-  "Large": {
-    "Tool": "yosys",
-    "Directive": "Default",
-    "Report": "large_default_flatten",
-    "YosysArgs": "-abc9 -flatten_before_abc"
-  },
-
-  "Large": {
-    "Tool": "vivado",
-    "Directive": "Default",
-    "Report": "large_default",
-    "YosysArgs": "-abc9"
-  }
-}
