+++ verilog/benchmarks_large/litex/soc_basesoc_minispartan6_rocket_ise.prj
@@ -1,4 +1,4 @@
-verilog work soc_basesoc_minispartan6_rocket.v.gz
+verilog work soc_basesoc_minispartan6_rocket.v
 verilog work rocket-litex-verilog/vsrc/plusarg_reader.v
 verilog work rocket-litex-verilog/generated-src/freechips.rocketchip.system.LitexConfig.v
 #verilog work rocket-litex-verilog/vsrc/EICG_wrapper.v
