+++ verilog/benchmarks_large/litex/soc_basesoc_minispartan6_vexriscv_ise.prj
@@ -1,2 +1,2 @@
-verilog work soc_basesoc_minispartan6_vexriscv.v.gz
+verilog work soc_basesoc_minispartan6_vexriscv.v
 verilog work VexRiscv-verilog/VexRiscv.v
